Add Blueprints to the application
