+++ .env.examples
@@ -2,6 +2,7 @@ source ./venv/bin/activate
 export FLASK_APP="run.py"
 export SECRET_KEY="<your secret phrase>"
 export DATABASE_URL=""
+export FLASK_CONFIG="<development>"
 export MS_TRANSLATOR_KEY=<paste-your-key-here>
 export ADMINS="["<paste-admin-email2>","<paste-admin-email2>"]"
