+++ .gitignore
@@ -5,3 +5,4 @@ venv/*
 app.db
 messages.pot
 .vscode/*
+logs/*
