+++ app/auth/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+auth_bp = Blueprint('auth', __name__,url_prefix="/auth")
+from . import routes
+
+
