+++ app/cli.py
@@ -1,43 +1,40 @@
 import os
 import click
 
-from app import app
 
+def register(app):
+    @app.cli.group()
+    def translate():
+        """Translation and localization commands."""
+        pass
 
-@app.cli.group()
-def translate():
-    """Translation and localization commands."""
-    pass
+    @translate.command()
+    @click.argument('lang')
+    def init(lang):
+        """Initialize a new language.
 
-@translate.command()
-@click.argument('lang')
-def init(lang):
-    """Initialize a new language.
+        Takes a language code argument where muliple language
+        translations exist. ie flask translate english
+        """
 
-    Takes a language code argument where muliple language
-    translations exist. ie flask translate english
-    """
+        if os.system('pybabel extract -F babel.cfg -k _l -o messages.pot .'):
+            raise RuntimeError('extract command failed')
+        if os.system(
+                'pybabel init -i messages.pot -d app/translations -l ' + lang):
+            raise RuntimeError('init command failed')
+        os.remove('messages.pot')
 
-    if os.system('pybabel extract -F babel.cfg -k _l -o messages.pot .'):
-        raise RuntimeError('extract command failed')
-    if os.system(
-            'pybabel init -i messages.pot -d app/translations -l ' + lang):
-        raise RuntimeError('init command failed')
-    os.remove('messages.pot')
+    @translate.command()
+    def update():
+        """Update all languages."""
+        if os.system('pybabel extract -F babel.cfg -k _l -o messages.pot .'):
+            raise RuntimeError('extract command failed')
+        if os.system('pybabel update -i messages.pot -d app/translations'):
+            raise RuntimeError('update command failed')
+        os.remove('messages.pot')
 
-
-@translate.command()
-def update():
-    """Update all languages."""
-    if os.system('pybabel extract -F babel.cfg -k _l -o messages.pot .'):
-        raise RuntimeError('extract command failed')
-    if os.system('pybabel update -i messages.pot -d app/translations'):
-        raise RuntimeError('update command failed')
-    os.remove('messages.pot')
-
-
-@translate.command()
-def compile():
-    """Compile all languages."""
-    if os.system('pybabel compile -d app/translations'):
-        raise RuntimeError('compile command failed')
+    @translate.command()
+    def compile():
+        """Compile all languages."""
+        if os.system('pybabel compile -d app/translations'):
+            raise RuntimeError('compile command failed')
