+++ app/email.py
@@ -1,10 +1,9 @@
 from threading import Thread
-from flask import render_template
+from flask import current_app
 from flask_mail import Message
-from flask_babel import _
+
 
 from app import mail
-from app import app
 
 
 def send_async_email(app, msg):
@@ -16,15 +15,6 @@ def send_email(subject, sender, recipients, text_body, html_body):
     msg = Message(subject, sender=sender, recipients=recipients)
     msg.body = text_body
     msg.html = html_body
-    Thread(target=send_async_email, args=(app, msg)).start()
+    Thread(target=send_async_email, args=(current_app, msg)).start()
 
 
-def send_password_reset_email(user):
-    token = user.get_reset_password_token()
-    send_email(_('[AKblog] Reset Your Password'),
-               sender=app.config['ADMINS'][0],
-               recipients=[user.email],
-               text_body=render_template('email/reset_password.txt',
-                                         user=user, token=token),
-               html_body=render_template('email/reset_password.html',
-                                         user=user, token=token))
