+++ app/error/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+error_bp= Blueprint('error', __name__)
+
+from . import handlers
