+++ app/error/handlers.py
@@ -1,13 +1,14 @@
 from flask import render_template
-from app import app, db
+from app import db
+from app.error import error_bp
 
 
-@app.errorhandler(404)
+@error_bp.errorhandler(404)
 def not_found_error(error):
     return render_template('errors/404.html'), 404
 
 
-@app.errorhandler(500)
+@error_bp.errorhandler(500)
 def internal_error(error):
     db.session.rollback()
     return render_template('errors/500.html'), 500
