+++ app/home/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+home_bp = Blueprint('home',__name__)
+from . import routes
