+++ app/home/routes.py
@@ -1,0 +1,143 @@
+from flask import render_template, flash, redirect, url_for, \
+    request, g, current_app
+from flask_login import current_user, login_required
+from datetime import datetime
+from flask_babel import _, get_locale
+from guess_language import guess_language
+
+
+from app import db
+from app.models import User, Post
+from . import home_bp
+from .forms import EditProfileForm, PostForm
+
+# from app.translate import translate
+
+
+@home_bp.before_request
+def before_request():
+    if current_user.is_authenticated:
+        current_user.last_seen = datetime.utcnow()
+        db.session.commit()
+    g.locale = str(get_locale())
+
+
+@home_bp.route('/', methods=['GET', 'POST'])
+@home_bp.route('/index', methods=['GET', 'POST'])
+@login_required
+def index():
+    postform = PostForm()
+
+    if postform.validate_on_submit():
+        language = guess_language(postform.post.data)
+
+        if language == 'UNKNOWN' or len(language) > 5:
+            language = ''
+
+        post = Post(body=postform.post.data, author=current_user, language=language)
+        db.session.add(post)
+        db.session.commit()
+        flash(_('Your post is now live!'))
+        return redirect(url_for('home.index'))
+
+    page = request.args.get('page', 1, type=int)
+    posts = current_user.followed_posts().paginate(page, current_app.config['POSTS_PER_PAGE'], False)
+
+    next_url = url_for('home.index', page=posts.next_num) \
+        if posts.has_next else None
+    prev_url = url_for('home.index', page=posts.prev_num) \
+        if posts.has_prev else None
+
+    return render_template('index.html', title=_('Home'), user=user,
+                           posts=posts.items, form=postform,
+                           next_url=next_url, prev_url=prev_url)
+
+
+@home_bp.route('/user/<username>')
+@login_required
+def user(username):
+    user = User.query.filter_by(username=username).first_or_404()
+    page = request.args.get('page', 1, type=int)
+    posts = user.posts.order_by(Post.timestamp.desc()).paginate(
+        page, current_app.config['POSTS_PER_PAGE'], False)
+    next_url = url_for('home.user', username=user.username, page=posts.next_num) \
+        if posts.has_next else None
+    prev_url = url_for('home.user', username=user.username, page=posts.prev_num) \
+        if posts.has_prev else None
+    return render_template('user.html', user=user, posts=posts.items,
+                           next_url=next_url, prev_url=prev_url)
+
+
+@home_bp.route('/edit_profile', methods=['GET', 'POST'])
+@login_required
+def edit_profile():
+    edit_profile_form = EditProfileForm(current_user.username)
+    if edit_profile_form.validate_on_submit():
+        current_user.username = edit_profile_form.username.data
+        current_user.about_me = edit_profile_form.about_me.data
+        db.session.commit()
+        flash(_('Your changes have been saved.'))
+        return redirect(url_for('home.edit_profile'))
+    elif request.method == 'GET':
+        edit_profile_form.username.data = current_user.username
+        edit_profile_form.about_me.data = current_user.about_me
+    return render_template('edit_profile.html', title=_('Edit Profile'),
+                           form=edit_profile_form)
+
+
+@home_bp.route('/follow/<username>')
+@login_required
+def follow(username):
+    user = User.query.filter_by(username=username).first()
+    if user is None:
+        flash(_('User %(username)s not found.', username=username))
+        return redirect(url_for('home.index'))
+    if user == current_user:
+        flash(_('You cannot follow yourself!'))
+        return redirect(url_for('home.user', username=username))
+    current_user.follow(user)
+    db.session.commit()
+    flash(_('You are following %(username)s!', username=username))
+    return redirect(url_for('home.user', username=username))
+
+
+@home_bp.route('/unfollow/<username>')
+@login_required
+def unfollow(username):
+    user = User.query.filter_by(username=username).first()
+    if user is None:
+        flash(_('User %(username)s not found.', username=username))
+        return redirect(url_for('home.index'))
+    if user == current_user:
+        flash(_('You cannot unfollow yourself!'))
+        return redirect(url_for('home.user', username=username))
+    current_user.unfollow(user)
+    db.session.commit()
+    flash(_('You are not following %(username)s.', username=username))
+    return redirect(url_for('home.user', username=username))
+
+
+@home_bp.route('/explore')
+@login_required
+def explore():
+    page = request.args.get('page', 1, type=int)
+
+    posts = Post.query.order_by(Post.timestamp.desc()).paginate(
+        page, current_app.config['POSTS_PER_PAGE'], False)
+
+    next_url = url_for('home.explore', page=posts.next_num) \
+        if posts.has_next else None
+
+    prev_url = url_for('home.explore', page=posts.prev_num) \
+        if posts.has_prev else None
+
+    return render_template('index.html', title=_('Explore'),posts=posts.items,
+                           next_url=next_url, prev_url=prev_url)
+
+
+# @app.route('/translate', methods=['POST'])
+# @login_required
+# def translate_text():
+#     return jsonify({'text': translate(request.form['text'],
+#                                       request.form['source_language'],
+#                                       request.form['dest_language'])})
