--- app/routes.py
@@ -1,220 +1,0 @@
-from flask import render_template, flash, redirect, url_for, \
-    request, g, jsonify
-from flask_login import current_user, login_user, logout_user, \
-    login_required
-from werkzeug.urls import url_parse
-from datetime import datetime
-from flask_babel import _, get_locale
-from guess_language import guess_language
-
-from app import app
-from app import db
-from app.forms import LoginForm, RegistrationForm, EditProfileForm, \
-    PostForm, ResetPasswordRequestForm, ResetPasswordForm
-
-from app.models import User, Post
-from app.email import send_password_reset_email
-#from app.translate import translate
-
-
-@app.before_request
-def before_request():
-    if current_user.is_authenticated:
-        current_user.last_seen = datetime.utcnow()
-        db.session.commit()
-    g.locale = str(get_locale())
-
-
-@app.route('/', methods=['GET', 'POST'])
-@app.route('/index', methods=['GET', 'POST'])
-@login_required
-def index():
-    postform = PostForm()
-
-    if postform.validate_on_submit():
-        language = guess_language(postform.post.data)
-
-        if language == 'UNKNOWN' or len(language) > 5:
-            language = ''
-
-        post = Post(body=postform.post.data, author=current_user, language=language)
-        db.session.add(post)
-        db.session.commit()
-        flash(_('Your post is now live!'))
-        return redirect(url_for('index'))
-
-    page = request.args.get('page', 1, type=int)
-    posts = current_user.followed_posts().paginate(page, app.config['POSTS_PER_PAGE'], False)
-
-    next_url = url_for('index', page=posts.next_num) \
-        if posts.has_next else None
-    prev_url = url_for('index', page=posts.prev_num) \
-        if posts.has_prev else None
-
-    return render_template('index.html', title=_('Home'), user=user,
-                           posts=posts.items, form=postform,
-                           next_url=next_url, prev_url=prev_url)
-
-
-@app.route('/user/<username>')
-@login_required
-def user(username):
-    user = User.query.filter_by(username=username).first_or_404()
-    page = request.args.get('page', 1, type=int)
-    posts = user.posts.order_by(Post.timestamp.desc()).paginate(
-        page, app.config['POSTS_PER_PAGE'], False)
-    next_url = url_for('user', username=user.username, page=posts.next_num) \
-        if posts.has_next else None
-    prev_url = url_for('user', username=user.username, page=posts.prev_num) \
-        if posts.has_prev else None
-    return render_template('user.html', user=user, posts=posts.items,
-                           next_url=next_url, prev_url=prev_url)
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-
-    login_form = LoginForm()
-
-    if login_form.validate_on_submit():
-        user = User.query.filter_by(username=login_form.username.data).first()
-        if user is None or not user.check_password(login_form.password.data):
-            flash(_('Invalid username or password'))
-            return redirect(url_for('login'))
-        login_user(user, remember=login_form.remember_me.data)
-
-        next_page = request.args.get('next')
-        if not next_page or url_parse(next_page).netloc != '':
-            next_page = url_for('index')
-        return redirect(next_page)
-
-    return render_template('login.html', title=_('Sign In'), form=login_form)
-
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for('index'))
-
-
-@app.route('/register', methods=['GET', 'POST'])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    reg_form = RegistrationForm()
-    if reg_form.validate_on_submit():
-        user = User(username=reg_form.username.data, email=reg_form.email.data)
-        user.set_password(reg_form.password.data)
-        db.session.add(user)
-        db.session.commit()
-        flash(_('Congratulations, you are now a registered user!'))
-        return redirect(url_for('login'))
-    return render_template('register.html', title=_('Register'), form=reg_form)
-
-
-@app.route('/edit_profile', methods=['GET', 'POST'])
-@login_required
-def edit_profile():
-    edit_profile_form = EditProfileForm(current_user.username)
-    if edit_profile_form.validate_on_submit():
-        current_user.username = edit_profile_form.username.data
-        current_user.about_me = edit_profile_form.about_me.data
-        db.session.commit()
-        flash(_('Your changes have been saved.'))
-        return redirect(url_for('edit_profile'))
-    elif request.method == 'GET':
-        edit_profile_form.username.data = current_user.username
-        edit_profile_form.about_me.data = current_user.about_me
-    return render_template('edit_profile.html', title=_('Edit Profile'),
-                           form=edit_profile_form)
-
-
-@app.route('/follow/<username>')
-@login_required
-def follow(username):
-    user = User.query.filter_by(username=username).first()
-    if user is None:
-        flash(_('User %(username)s not found.', username=username))
-        return redirect(url_for('index'))
-    if user == current_user:
-        flash(_('You cannot follow yourself!'))
-        return redirect(url_for('user', username=username))
-    current_user.follow(user)
-    db.session.commit()
-    flash(_('You are following %(username)s!', username=username))
-    return redirect(url_for('user', username=username))
-
-
-@app.route('/unfollow/<username>')
-@login_required
-def unfollow(username):
-    user = User.query.filter_by(username=username).first()
-    if user is None:
-        flash(_('User %(username)s not found.', username=username))
-        return redirect(url_for('index'))
-    if user == current_user:
-        flash(_('You cannot unfollow yourself!'))
-        return redirect(url_for('user', username=username))
-    current_user.unfollow(user)
-    db.session.commit()
-    flash(_('You are not following %(username)s.', username=username))
-    return redirect(url_for('user', username=username))
-
-
-@app.route('/explore')
-@login_required
-def explore():
-    page = request.args.get('page', 1, type=int)
-
-    posts = Post.query.order_by(Post.timestamp.desc()).paginate(
-        page, app.config['POSTS_PER_PAGE'], False)
-
-    next_url = url_for('explore', page=posts.next_num) \
-        if posts.has_next else None
-
-    prev_url = url_for('explore', page=posts.prev_num) \
-        if posts.has_prev else None
-
-    return render_template('index.html', title=_('Explore'),posts=posts.items,
-                           next_url=next_url, prev_url=prev_url)
-
-
-@app.route('/reset_password_request', methods=['GET', 'POST'])
-def reset_password_request():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    reset_password_request_form = ResetPasswordRequestForm()
-    if reset_password_request_form.validate_on_submit():
-        user = User.query.filter_by(email=reset_password_request_form.email.data).first()
-        if user:
-            send_password_reset_email(user)
-        flash(_('Check your email for the instructions to reset your password'))
-        return redirect(url_for('login'))
-    return render_template('reset_password_request.html', title=_('Reset Password'),
-                           form=reset_password_request_form)
-
-
-@app.route('/reset_password/<token>', methods=['GET', 'POST'])
-def reset_password(token):
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    user = User.verify_reset_password_token(token)
-    if not user:
-        return redirect(url_for('index'))
-    reset_password_form = ResetPasswordForm()
-    if reset_password_form.validate_on_submit():
-        user.set_password(reset_password_form.password.data)
-        db.session.commit()
-        flash(_('Your password has been reset.'))
-        return redirect(url_for('login'))
-    return render_template('reset_password.html', form=reset_password_form)
-
-
-# @app.route('/translate', methods=['POST'])
-# @login_required
-# def translate_text():
-#     return jsonify({'text': translate(request.form['text'],
-#                                       request.form['source_language'],
-#                                       request.form['dest_language'])})
