+++ app/templates/_post.html
@@ -1,13 +1,13 @@
 <table class="table table-hover">
     <tr>
         <td width="70px">
-            <a href="{{ url_for('user', username=post.author.username) }}">
+            <a href="{{ url_for('home.user', username=post.author.username) }}">
                 <img src="{{ post.author.avatar(70) }}"/>
             </a>
         </td>
         <td>
             {% set user_link %}
-            <a href="{{ url_for('user', username=post.author.username) }}">
+            <a href="{{ url_for('home.user', username=post.author.username) }}">
                 {{ post.author.username }}
             </a>
             {% endset %}
