+++ app/templates/errors/500.html
@@ -3,5 +3,5 @@
 {% block app_content %}
     <h1>{{ _('An unexpected error has occurred') }}</h1>
     <p>{{ _('The administrator has been notified. Sorry for the inconvenience!') }}</p>
-    <p><a href="{{ url_for('index') }}">{{ _('Back') }}</a></p>
+    <p><a href="{{ url_for('home.index') }}">{{ _('Back') }}</a></p>
 {% endblock %}
