+++ app/templates/user.html
@@ -12,11 +12,11 @@
                 {% endif %}
                 <p>{{ _('%(count)d followers', count=user.followers.count()) }}, {{ _('%(count)d following', count=user.followed.count()) }}</p>
                 {% if user == current_user %}
-                <p><a href="{{ url_for('edit_profile') }}">{{ _('Edit your profile') }}</a></p>
+                <p><a href="{{ url_for('home.edit_profile') }}">{{ _('Edit your profile') }}</a></p>
                 {% elif not current_user.is_following(user) %}
-                <p><a href="{{ url_for('follow', username=user.username) }}">{{ _('Follow') }}</a></p>
+                <p><a href="{{ url_for('home.follow', username=user.username) }}">{{ _('Follow') }}</a></p>
                 {% else %}
-                <p><a href="{{ url_for('unfollow', username=user.username) }}">{{ _('Unfollow') }}</a></p>
+                <p><a href="{{ url_for('home.unfollow', username=user.username) }}">{{ _('Unfollow') }}</a></p>
                 {% endif %}
             </td>
         </tr>
