+++ requirements.txt
@@ -1,6 +1,9 @@
 alembic==1.0.10
+astroid==2.2.5
 Babel==2.7.0
 blinker==1.4
+certifi==2019.3.9
+chardet==3.0.4
 Click==7.0
 dominate==2.3.5
 Flask==1.0.3
@@ -12,17 +15,27 @@ Flask-Migrate==2.5.2
 Flask-Moment==0.7.0
 Flask-SQLAlchemy==2.4.0
 Flask-WTF==0.14.2
+guess-language-spirit==0.5.3
+idna==2.8
+isort==4.3.20
 itsdangerous==1.1.0
 Jinja2==2.10.1
+lazy-object-proxy==1.4.1
 Mako==1.0.12
 MarkupSafe==1.1.1
+mccabe==0.6.1
 nose==1.3.7
 PyJWT==1.7.1
+pylint==2.3.1
 python-dateutil==2.8.0
 python-editor==1.0.4
 pytz==2019.1
+requests==2.22.0
 six==1.12.0
 SQLAlchemy==1.3.4
+typed-ast==1.4.0
+urllib3==1.25.3
 visitor==0.1.3
 Werkzeug==0.15.4
+wrapt==1.11.1
 WTForms==2.2.1
