+++ run.py
@@ -1,5 +1,10 @@
-from app import app, cli
+import os
+from app import create_app, cli
 
+app_environment = os.getenv('FLASK_CONFIG') or 'production'
+app = create_app(app_environment)
+
+cli.register(app)
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run()
